Move iaphandler out of "platforms"
