+++ src/iaphandler.cpp
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-#include "platforms/iaphandler.h"
+#include "iaphandler.h"
 #include "constants.h"
 #include "leakdetector.h"
 #include "logger.h"
